workspace/7_framework/fastapi/day11/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.23
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
pydantic_settings
pydantic[email]
